Changed migration

+++ database/migrations/2019_06_07_210018_create_discord__accounts_table.php
@@ -14,7 +14,7 @@ class CreateDiscordAccountsTable extends Migration
     public function up()
     {
         Schema::create('discord__accounts', function (Blueprint $table) {
-            $table->unsignedInteger('user_id')->unique();
+            $table->unsignedBigInteger('user_id')->unique();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->unsignedBigInteger('id')->unique();
         });
